Reorganize edwards curves file structure
