+++ math/src/elliptic_curve/edwards/curves/mod.rs
@@ -1,0 +1,1 @@
+pub mod tiny_jub_jub;
