+++ math/src/elliptic_curve/edwards/mod.rs
@@ -1,2 +1,3 @@
+pub mod curves;
 pub mod point;
 pub mod traits;
